+++ services/src/main/java/by/mozgo/craps/services/game/RollResult.java
@@ -7,19 +7,19 @@ package by.mozgo.craps.services.game;
  *
  */
 public class RollResult {
-    private int dice1;
-    private int dice2;
+    private byte dice1;
+    private byte dice2;
 
-    public RollResult(int dice1, int dice2){
+    public RollResult(byte dice1, byte dice2) {
         this.dice1 = dice1;
         this.dice2 = dice2;
     }
 
-    public int getDice1() {
+    public byte getDice1() {
         return dice1;
     }
 
-    public int getDice2() {
+    public byte getDice2() {
         return dice2;
     }
 }
